requirements.txt
python-telegram-bot==21.6
pandas==2.2.2
loguru==0.7.2
APScheduler==3.10.4
requests==2.32.3
websockets==12.0
python-dotenv==1.0.1
sqlAlchemy==2.0.32
PyYAML==6.0.1
pytz==2024.1
